AWS: address Direct Connect review comments
Switch from pre-installed static routes to BGP redistribution between
TGW and DXGW, matching AWS's actual behavior more closely. Routes now
appear on the TGW only when the BGP session is up — a misconfigured
or down VIF no longer leaves stale "reachable" routes on the TGW.

DXGW redesign:
- Originate each association's allowedPrefixesToDirectConnectGateway as
  static null-routes on the DXGW; advertise these (and only these) to
  the on-prem customer via BGP, matching AWS's documented behavior of
  advertising the allowed-prefix list itself rather than the underlying
  per-VPC routes.
- Per-association TGW-export policy filters routes received from on-prem
  to those within or equal to one of that association's allowed prefixes,
  matching AWS's description of the prefix list as an inbound filter.

TGW side:
- Connect TGW VRF to DXGW default VRF via BGP unnumbered over the existing
  link-local interfaces.
- Per-route-table import policy on the DXGW peer tags incoming BGP routes
  with the DX-propagated admin distance, so DX wins over VPN (default BGP
  admin) for the same prefix in TGW route selection.
- Route-table-aware: each DX attachment is associated with exactly one TGW
  route table, and BGP peering is established only in that VRF.
  VPC-to-DXGW route advertisement falls out naturally from BGP exporting
  the per-VRF RIB, fixing the multi-route-table leak from the previous
  static-route-installation path.

Other cleanups:
- Drop the parsed-but-unused _routeFilterPrefixes field from
  DirectConnectVirtualInterface; for Transit VIFs the operative filter is
  the association's allowed prefixes, and the dead field would silently
  do the wrong thing if Private/Public VIF support is added later.
- Replace tautological constant-comparison test
  (testRoutePreferenceOrdering) with one that inspects the TGW's RIB and
  asserts that the BGP route propagated from the DXGW carries the
  DX-propagated admin distance.
- Adjust the test on-prem router config (remove overly aggressive
  10.0.0.0/8 null-route, advertise a more specific 10.10.0.0/16 instead);
  the previous null-route was masking the DXGW's BGP advertisement.
- Make DIRECT_CONNECT_PROPAGATED_ROUTE_ADMIN a long to match
  AbstractRoute.getAdministrativeCost.

Prompt:
```
Read the review comments on the issue. What do you think? I haven't
thought deeply at the moment.
```

Followed up by: "nah, do all of them" — addresses all four review
comments on PR-equivalent commit 923c204e.

+++ projects/batfish/src/main/java/org/batfish/representation/aws/Route.java
@@ -49,7 +49,7 @@ abstract class Route implements Serializable {
   /**
    * Admin distance for routes propagated from Direct Connect (preferred over VPN but not static)
    */
-  static final int DIRECT_CONNECT_PROPAGATED_ROUTE_ADMIN = 10;
+  static final long DIRECT_CONNECT_PROPAGATED_ROUTE_ADMIN = 10L;
 
   static final int DEFAULT_STATIC_ROUTE_COST = 0;
